wip datasets via entry points
