+++ pyproject.toml
@@ -41,3 +41,7 @@ CI = "https://github.com/terrier-org/pyterrier/actions"
 [project.entry-points."pyterrier.java.init"]
 "pyterrier.java" = "pyterrier.java:CoreJavaInit"
 "pyterrier.terrier.java" = "pyterrier.terrier.java:TerrierJavaInit"
+
+[project.entry-points."pyterrier.dataset_provider"]
+"builtin" = "pyterrier.datasets._builtin:BuiltinDatasetProvider"
+"irds" = "pyterrier.datasets._irds:IRDSDatasetProvider"
